Now towers shoot projectiles when enemies are in their range and enemies get damaged when a projectile enters their hitbox

+++ src/main/java/com/mfarioli/JavaTD/Helpers/Constants.java
@@ -88,7 +88,7 @@ public class Constants {
             return "";
         }
 
-        public static float getStartingDamage(int towerType) {
+        public static int getStartingDamage(int towerType) {
             switch (towerType) {
                 case CANNON -> {
                     return 25;
